move getAdditionalServices to helpers
